add auto + bind to resource
