+++ packages/config-transformer/packages/format-switcher/tests/Converter/ConfigFormatConverter/FixtureYamlToPhp/normal/constant.yaml
@@ -2,6 +2,8 @@ parameters:
     indentation: 'spaces'
     another: PHP_EOL
     line_ending: !php/const PHP_EOL
+    older_line_ending: !php/const:PHP_EOL
+    older_line_ending_with_space: !php/const: PHP_EOL
 -----
 <?php
 
@@ -15,4 +17,8 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $parameters->set('another', 'PHP_EOL');
 
     $parameters->set('line_ending', PHP_EOL);
+
+    $parameters->set('older_line_ending', PHP_EOL);
+
+    $parameters->set('older_line_ending_with_space', PHP_EOL);
 };
